Ajustes integracao front e back tela de produtos - requisito 13 nao passa

+++ back-end/src/database/migrations/20230120200606-products.js
@@ -16,7 +16,7 @@ module.exports = {
       },
       price: {
         allowNull: false,
-        type: Sequelize.FLOAT
+        type: Sequelize.DECIMAL(4,2) // troca Float por Decimal(4,2)
       },
       url_image: {
         allowNull: false,
